backend/requirements.txt
# FastAPI и веб-сервер
fastapi==0.104.1
uvicorn[standard]==0.24.0

# База данных
sqlalchemy==2.0.23
mysqlclient==2.2.0
alembic==1.12.1

# Авторизация и безопасность
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Валидация данных
pydantic==2.5.0
pydantic-settings==2.1.0

# Утилиты
python-dotenv==1.0.0
email-validator==2.1.0
requests==2.31.0

# Разработка и тестирование
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# CORS для фронтенда
fastapi-cors==0.0.6

# Работа с файлами
aiofiles==23.2.1
Pillow==10.1.0

# Логирование
loguru==0.7.2